+++ bob/src/hw_metrics_collector.rs
@@ -1,12 +1,12 @@
 use crate::prelude::*;
 use bob_common::metrics::{
-    BOB_RAM, CPU_LOAD, DESCRIPTORS_AMOUNT, FREE_RAM, FREE_SPACE, TOTAL_RAM, TOTAL_SPACE, USED_RAM,
-    USED_SPACE,
+    BOB_RAM, CPU_IOWAIT, CPU_LOAD, DESCRIPTORS_AMOUNT, FREE_RAM, FREE_SPACE, HW_DISKS_FOLDER,
+    TOTAL_RAM, TOTAL_SPACE, USED_RAM, USED_SPACE,
 };
 use libc::statvfs;
 use std::os::unix::fs::MetadataExt;
 use std::path::{Path, PathBuf};
-use std::process;
+use std::process::{self, Command};
 use sysinfo::{DiskExt, ProcessExt, RefreshKind, System, SystemExt};
 
 const DESCRS_DIR: &str = "/proc/self/fd/";
@@ -66,6 +66,8 @@ impl HWMetricsCollector {
         let mut interval = interval(t);
         let mut sys = System::new_all();
         let mut dcounter = DescrCounter::new();
+        let mut cpu_s_c = CPUStatCollector::new();
+        let mut disk_s_c = DiskStatCollector::new(&disks);
         let total_mem = kb_to_mb(sys.total_memory());
         gauge!(TOTAL_RAM, total_mem as f64);
         debug!("total mem in mb: {}", total_mem);
@@ -73,23 +75,42 @@ impl HWMetricsCollector {
 
         loop {
             interval.tick().await;
+
             sys.refresh_specifics(
                 RefreshKind::new()
                     .with_processes()
                     .with_disks()
                     .with_memory(),
             );
-            let proc = sys.process(pid).expect("Can't get process stat descriptor");
+
+            match cpu_s_c.iowait() {
+                Ok(iowait) => {
+                    gauge!(CPU_IOWAIT, iowait);
+                },
+                Err(CommandError::Primary(e)) => {
+                    warn!("Error while collecting cpu iowait: {}", e);
+                },
+                Err(CommandError::Unavailable) => (),
+            }
+
+            if let Some(proc) = sys.process(pid) {
+                gauge!(CPU_LOAD, proc.cpu_usage() as f64);
+                let bob_ram = kb_to_mb(proc.memory());
+                gauge!(BOB_RAM, bob_ram as f64);
+            } else {
+                debug!("Can't get process stat descriptor");
+            }
 
             let _ = Self::update_space_metrics_from_disks(&disks);
             let used_mem = kb_to_mb(sys.used_memory());
             debug!("used mem in mb: {}", used_mem);
             gauge!(USED_RAM, used_mem as f64);
             gauge!(FREE_RAM, (total_mem - used_mem) as f64);
-            let bob_ram = kb_to_mb(proc.memory());
-            gauge!(BOB_RAM, bob_ram as f64);
             gauge!(DESCRIPTORS_AMOUNT, dcounter.descr_amount() as f64);
-            gauge!(CPU_LOAD, proc.cpu_usage() as f64);
+
+            if let Err(CommandError::Primary(e)) = disk_s_c.collect_and_send_metrics() {
+                warn!("Error while collecting stats of disks: {}", e);
+            }
         }
     }
 
@@ -125,6 +146,7 @@ impl HWMetricsCollector {
     // refreshed, if I clone them
     // NOTE: HashMap contains only needed mount points of used disks, so it won't be really big,
     // but maybe it's more efficient to store disks (instead of mount_points) and update them one by one
+
     fn space(disks: &HashMap<PathBuf, String>) -> DiskSpaceMetrics {
         let mut total = 0;
         let mut used = 0;
@@ -148,6 +170,165 @@ impl HWMetricsCollector {
             total_space: total,
             used_space: used,
             free_space: free,
+        }
+    }
+}
+
+enum CommandError {
+    Unavailable,
+    Primary(String),
+}
+struct CPUStatCollector {
+    iostat_avl: bool,
+}
+
+impl CPUStatCollector {
+    fn new() -> CPUStatCollector {
+        CPUStatCollector { iostat_avl: true }
+    }
+
+    fn parse_iostat_result(iostat_str: &str) -> Result<f64, std::num::ParseFloatError> {
+        iostat_str.replace(',', ".").parse()
+    }
+
+    fn iowait(&mut self) -> Result<f64, CommandError> {
+        if !self.iostat_avl {
+            return Err(CommandError::Unavailable);
+        }
+
+        match parse_command_output(Command::new("iostat").arg("-c")) {
+            Ok(output) => {
+                // find avg-cpu headers line
+                let mut lines = output
+                    .lines()
+                    .skip_while(|line| line.find("avg-cpu:").is_none());
+                // find iowait column on the next line
+                if let Some(line) = lines.next() {
+                    let mut values = line.split_whitespace();
+                    // iowait is in 3d column
+                    if let Some(iowait_str) = values.nth(3) {
+                        if let Ok(iowait) = Self::parse_iostat_result(iowait_str) {
+                            return Ok(iowait);
+                        }
+                    }
+                }
+            }
+            Err(e) => {
+                self.iostat_avl = false;
+                return Err(CommandError::Primary(e));
+            }
+        }
+        self.iostat_avl = false;
+        Err(CommandError::Primary("Failed to get iowait from iostat".to_string()))
+    }
+}
+
+struct DiskStatCollector {
+    iostat_avl: bool,
+    disk_metric_data: HashMap<String, String>,
+}
+
+impl DiskStatCollector {
+    fn new(disks: &HashMap<PathBuf, String>) -> Self {
+        let mut disk_metric_data = HashMap::new();
+        for (path, disk_name) in disks {
+            let path_str = path.as_os_str().to_str().unwrap();
+            if let Ok(dev_name) = Self::dev_name(path_str) {
+                let metric_prefix = format!("{}.{}", HW_DISKS_FOLDER, disk_name);
+                disk_metric_data.insert(dev_name, metric_prefix);
+            } else {
+                warn!("Device name of disk {} is unknown", disk_name);
+            }
+        }
+
+        DiskStatCollector {
+            iostat_avl: true,
+            disk_metric_data,
+        }
+    }
+
+    fn collect_and_send_metrics(&mut self) -> Result<(), CommandError> {
+        let iostat_lines = self.collect_iostat()?;
+        for i in iostat_lines {
+            let lsp: Vec<&str> = i.split_whitespace().collect();
+            if lsp.len() == 21 {
+                if let Some(metric_prefix) = self.disk_metric_data.get(lsp[0]) {
+                    if let (Ok(rs), Ok(ws)) = (
+                        // readops per s count is in 1st column
+                        Self::parse_iostat_result(lsp[1]),
+                        // writeops per s count is in 7th column
+                        Self::parse_iostat_result(lsp[7]),
+                    ) {
+                        let iops = rs + ws;
+
+                        let gauge_name = format!("{}_iops", metric_prefix);
+                        gauge!(gauge_name, iops);
+                    } else {
+                        return Err(CommandError::Primary("Can't parse iostat output into float".to_string()));
+                    }
+
+                    if let (Ok(rwait), Ok(wwait)) = (
+                        // readops wait time is in 5th column
+                        Self::parse_iostat_result(lsp[5]),
+                        // writeops wait time is in 11th column
+                        Self::parse_iostat_result(lsp[11]),
+                    ) {
+                        let iowait = rwait + wwait;
+
+                        let gauge_name = format!("{}_iowait", metric_prefix);
+                        gauge!(gauge_name, iowait);
+                    } else {
+                        return Err(CommandError::Primary("Can't parse iostat output into float".to_string()));
+                    }
+                }
+            } else if lsp.len() > 0 {
+                return Err(CommandError::Primary("iostat output format changed, update this code".to_string()));
+            }
+        }
+        Ok(())
+    }
+
+    fn dev_name(disk_path: &str) -> Result<String, String> {
+        let output = parse_command_output(Command::new("df").arg(disk_path))?;
+
+        let mut lines = output.lines();
+        lines.next(); // skip headers
+        if let Some(line) = lines.next() {
+            if let Some(raw_dev_name) = line.split_whitespace().next() {
+                if let (Some(slash_ind), Some(non_digit_ind)) = (
+                    // find where /dev/ ends
+                    raw_dev_name.rfind('/'),
+                    // find where partition digits start
+                    raw_dev_name.rfind(|c: char| !c.is_digit(10)),
+                ) {
+                    return Ok(raw_dev_name[slash_ind + 1..=non_digit_ind].to_string());
+                }
+            }
+        }
+
+        Err("Error occured in dev_name getter".to_string())
+    }
+
+    fn parse_iostat_result(iostat_str: &str) -> Result<f64, std::num::ParseFloatError> {
+        iostat_str.replace(',', ".").parse()
+    }
+
+    fn collect_iostat(&mut self) -> Result<Vec<String>, CommandError> {
+        if !self.iostat_avl {
+            return Err(CommandError::Unavailable);
+        }
+
+        match parse_command_output(Command::new("iostat").arg("-dx")) {
+            Ok(output) => {
+                let line_iter = output.lines();
+                // skip headers
+                let line_iter = line_iter.skip(3);
+                Ok(line_iter.map(|line| line.to_string()).collect())
+            }
+            Err(e) => {
+                self.iostat_avl = false;
+                Err(CommandError::Primary(e))
+            }
         }
     }
 }
@@ -264,4 +445,30 @@ fn bytes_to_mb(bytes: u64) -> u64 {
 
 fn kb_to_mb(kbs: u64) -> u64 {
     kbs / 1024
+}
+
+fn parse_command_output(command: &mut Command) -> Result<String, String> {
+    let output = command.output();
+    let program = command.get_program().to_str().unwrap();
+    match output {
+        Ok(output) => match (output.status.success(), String::from_utf8(output.stdout)) {
+            (true, Ok(out)) => Ok(out),
+            (false, _) => {
+                if let Ok(e) = String::from_utf8(output.stderr) {
+                    let error = format!("Command {} finished with error: {}", program, e);
+                    Err(error)
+                } else {
+                    Err("Command finished with error".to_string())
+                }
+            }
+            (_, Err(e)) => {
+                let error = format!("Can not convert output of {} into string: {}", program, e);
+                Err(error)
+            }
+        },
+        Err(e) => {
+            let error = format!("Failed to execute command {}: {}", program, e);
+            Err(error)
+        }
+    }
 }